pyproject.toml
[tool.poetry]
name = "book-shop-drf-p23"
version = "0.1.0"
description = ""
authors = ["Hasanboy09 <sanoqulovhasan113@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django-mptt = "^0.16.0"
django = "5.1.2"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
drf-spectacular = "^0.27.2"
pillow = "^10.4.0"
djangorestframework-simplejwt = "^5.3.1"
celery = {extras = ["redis"], version = "^5.4.0"}
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
